- Every CRUD Operation warns (in log) if a publisher/game is being referenced that does not exist

+++ src/main/java/com/gamesproject/controller/GameController.java
@@ -1,7 +1,7 @@
 package com.gamesproject.controller;
 
 import com.gamesproject.dao.GamesDao;
-import com.gamesproject.dao.GamesDao;
+import com.gamesproject.dao.PublisherDao;
 import com.gamesproject.dto.Games;
 import com.gamesproject.dto.Publisher;
 import org.slf4j.Logger;
@@ -11,7 +11,6 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.context.request.NativeWebRequest;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -34,36 +33,73 @@ public class GameController extends AbstractController implements GamesApi{
         GamesDao gamesDao = new GamesDao(jdbcTemplate);
         logger.info("Successfully selected all Games");
         return getRespond(gamesDao.selectAll());
+
     }
 
     @Override
     public ResponseEntity<Games> gamesIdGet(Integer id) {
         GamesDao gamesDao = new GamesDao(jdbcTemplate);
         Optional<Games> games = gamesDao.selectById(id);
-        logger.info("Successfully selected a Game");
-        return getRespond(games.orElse(null));
+        if (gamesDao.selectById(id).isEmpty()){
+            logger.warn("Selected ID does not exist");
+            return errorResponse();
+        }
+        else {
+            logger.info("Successfully selected a Game");
+            return getRespond(games.orElse(null));
+        }
     }
 
 
     @Override
     public ResponseEntity<Integer> gamesIdDelete(Integer id) {
         GamesDao gamesDao = new GamesDao(jdbcTemplate);
-        logger.info("Successfully deleted a Game");
-        return deleteRespond(gamesDao.deleteById(id));
+        if (gamesDao.selectById(id).isEmpty()){
+            logger.warn("Selected ID does not exist");
+            return errorResponse();
+        }
+        else {
+            logger.info("Successfully deleted a Game");
+            return deleteRespond(gamesDao.deleteById(id));
+        }
     }
 
     @Override
     public ResponseEntity<Void> gamesPost(Games game) {
-        GamesDao gamesDao = new GamesDao(jdbcTemplate);
-        gamesDao.insert(game);
-        logger.info("Successfully created a Game");
-        return postRespond();
+        try {
+            GamesDao gamesDao = new GamesDao(jdbcTemplate);
+            PublisherDao publisherDao = new PublisherDao(jdbcTemplate);
+            if (publisherDao.selectById(game.getPublisherId()).isEmpty()){
+                logger.warn("Selected publisher ID does not exist");
+                return errorResponse();
+            }
+            gamesDao.insert(game);
+            logger.info("Successfully created a Game");
+            return postRespond();
+        }
+        catch (Exception ex){
+            logger.warn(ex.getMessage());
+            return errorResponse();
+        }
     }
 
     @Override
     public ResponseEntity<Integer> gamesPut(Games game) {
-        GamesDao gamesDao = new GamesDao(jdbcTemplate);
-        logger.info("Successfully updated a Game");
-        return putRespond(gamesDao.update(game));
+        try {
+            GamesDao gamesDao = new GamesDao(jdbcTemplate);
+            PublisherDao publisherDao = new PublisherDao(jdbcTemplate);
+            logger.info("Successfully updated a Game");
+            if (publisherDao.selectById(game.getId()).isEmpty()){
+                logger.warn("Selected game does not exist");
+                return errorResponse();
+            }
+            else {
+                return putRespond(gamesDao.update(game));
+            }
+        }
+        catch (Exception ex){
+            logger.warn(ex.getMessage());
+            return errorResponse();
+        }
     }
 }
