+++ src/main/java/com/gamesproject/controller/PublisherController.java
@@ -1,6 +1,5 @@
 package com.gamesproject.controller;
 
-import com.gamesproject.dao.GamesDao;
 import com.gamesproject.dao.PublisherDao;
 import com.gamesproject.dto.Publisher;
 import org.slf4j.Logger;
@@ -39,31 +38,68 @@ public class PublisherController extends AbstractController implements Publisher
     public ResponseEntity<Publisher> publisherIdGet(Integer id) {
         PublisherDao publisherDao = new PublisherDao(jdbcTemplate);
         Optional<Publisher> publisher = publisherDao.selectById(id);
-        logger.info("Successfully selected a Publisher");
-        return getRespond(publisher.orElse(null));
+        if (publisherDao.selectById(id).isEmpty()){
+            logger.warn("Selected publisher does not exist");
+            return errorResponse();
+        }
+        else {
+            logger.info("Successfully selected a Publisher");
+            return getRespond(publisher.orElse(null));
+        }
+
     }
 
 
     @Override
     public ResponseEntity<Integer> publisherIdDelete(Integer id) {
-        PublisherDao publisherDao = new PublisherDao(jdbcTemplate);
-        logger.info("Successfully deleted a Publisher");
-        return deleteRespond(publisherDao.deleteById(id));
+            PublisherDao publisherDao = new PublisherDao(jdbcTemplate);
+            logger.info("Successfully deleted a Publisher");
+            if (publisherDao.selectById(id).isEmpty()){
+                logger.warn("Selected publisher does not exist");
+                return errorResponse();
+            }
+            else {
+                try {
+                    return deleteRespond(publisherDao.deleteById(id));
+                }
+                catch (Exception ex){
+                    logger.warn("Selected publisher cannot be deleted because it is referenced in a game");
+                    return errorResponse();
+                }
+            }
     }
 
     @Override
     public ResponseEntity<Void> publisherPost(Publisher publisher) {
-        PublisherDao publisherDao = new PublisherDao(jdbcTemplate);
-        publisherDao.insert(publisher);
-        logger.info("Successfully created a Publisher");
-        return postRespond();
+        try {
+            PublisherDao publisherDao = new PublisherDao(jdbcTemplate);
+            publisherDao.insert(publisher);
+                logger.info("Successfully created a Publisher");
+                return postRespond();
+        }
+        catch (Exception ex){
+            logger.warn(ex.getMessage());
+            return errorResponse();
+        }
     }
 
     @Override
 
     public ResponseEntity<Integer> publisherPut(Publisher publisher) {
-        PublisherDao publisherDao = new PublisherDao(jdbcTemplate);
-        logger.info("Successfully updated a Publisher");
-        return putRespond(publisherDao.update(publisher));
+        try {
+            PublisherDao publisherDao = new PublisherDao(jdbcTemplate);
+            if (publisherDao.selectById(publisher.getId()).isEmpty()){
+                logger.warn("Selected publisher does not exist");
+                return errorResponse();
+            }
+            else {
+                logger.info("Successfully updated a Publisher");
+                return putRespond(publisherDao.update(publisher));
+            }
+        }
+        catch (Exception ex){
+            logger.warn(ex.getMessage());
+            return errorResponse();
+        }
     }
 }
